requirements.txt
aiohttp
cryptography
discord.py
